Fix class and module docs
Class and module documentation was being generated, however it
was not linked anywhere that was useful on the docs pages.
This commit cleans up some code that was required with older
versions of sphinx.

This commit puts all class/module documentation on the same
page as the class/module itself.

Change-Id: I864f099f90527405c7b8ac895236b418149a9d99